Coupon redeem added

+++ blog/views.py
@@ -21,13 +21,13 @@ class PostCreateAndListView(viewsets.ModelViewSet):
     
     def get_queryset(self):
         try:
-            queryset = self.queryset.filter(user=self.request.user)
+            queryset = self.queryset
         except Post.DoesNotExist:
             return Response({"DOESNT_EXIST": "No object in post list"}, status=402)
         title = self.request.query_params.get('title', None)
         # print(title)
         if title is not None:
-            queryset = queryset.filter(Q(title__icontains=title))
+            queryset = queryset.filter(Q(title__icontains=title) or Q(body__icontains=title))
 
         return queryset
 
@@ -54,20 +54,7 @@ class PostCreateAndListView(viewsets.ModelViewSet):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class FilterPost(viewsets.ModelViewSet):
-    queryset = Post.objects.all()
-    serializer_class = PostSerializer
-    permission_classes = [permissions.IsAuthenticated, IsCustomer]
 
-    def get_queryset(self):
-        queryset = self.queryset
-        print(queryset)
-        title = self.request.query_params.get('title', None)
-        print(title)
-        if title is not None:
-            queryset = queryset.filter(Q(title__icontains=title))
-
-        return queryset
 
 class BlogDelUpdate(APIView):
     queryset = Post.objects.all()
@@ -79,13 +66,17 @@ class BlogDelUpdate(APIView):
             return Post.objects.get(id=pk)
         except Post.DoesNotExist:
             raise Http404
+
     def get(self, request, pk, format=None):
         snippet = self.get_object(pk)
+    
         serializer = self.serializer_class(snippet)
         return Response(serializer.data)
 
     def delete(self, request, pk, format=None):
         queryset = self.get_object(pk)
+        if queryset.user is not request.user:
+            return Response({"DOES_NOT_EXIST": "User Does not have delete permission"}, status=400)
         queryset.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
